Remove debug prints and add error handling in git_provider
Removed print statements used for debugging in git_provider.py. An additional error handling mechanism was implemented for writing secrets. Now instead of just returning an 'OK' status, the function also checks if the secret storing process is successful or not and throws an internal server error if it fails.