Add upload codebase mutation and enhance logging
An uploadCodebase mutation was added to the 'legacy/mutations.py' file to allow the uploading of codebases. Additional logging was also inserted into the 'onboarding_event_handler' for the sake of better tracking and debugging. Codebase metadata is included in the upload process to offer more context about the file being uploaded.