Add MODAL_ENVIRONMENT configuration and organization_id
A new configuration variable, MODAL_ENVIRONMENT, is added in the backend core configuration. This variable is subsequently used for setting up the environment in a newly added function for PDF summary creation. Additionally, a new field, organization_id, has been added in the authentication schema. This commit also includes changes to the develop.yaml to allow workflow to run on an additional branch, "eric/trigger-pdf-summarization".