PE-921 implement organization and user self management endpoints (#624)
Also addresses PE-915 See those tickets for more details and examples if required. These endpoints (mostly) just proxy to and from Auth0. In a couple of places, we do some additional data transformation to make it more convenient for us.

Introduces unit tests in auth0_service_test.py. These tests do not require anything else to be running, run quickly, and verify that tokens without the organization:management permission fail. I think this is a productive and useful pattern to use in lieu of full-on reproducible e2e environments and all of infrastructure supporting those entails.