PE-2361: Add 'substitutions' to AutoDoc config
This change introduces the concept of 'substitutions' in the AutoDoc
config files.  This allows the option of defining an arbitrary number of
key/value pairs in the TOML config file.  When using the keys as
place holders in section instructions and/or content_structure
definitions, the system will perform a text substitution replacing each
key with the value before generating the document.