PE-2648: Add AutoTOML feature and accompanying CLI
This change introduces a new feature that allows users to automatically
create a AutoDocs TOML configuration file from a given set of sources
and document goal.  Additionally, users can supply a set of sources
and an existing TOML file to extend with additional sections
automatically.

The feature is exposed through a local CLI and not yet integrated in to
the product.