PE-2649: Refactor AutoDoc TOML config file
This change defines default values for many of the keys in the AutoDoc
TOML config file allowing users to specify the only the bare minimum
values.  This prevents users from having to specify technical details in
the config and allows them to focus only on defining the document
content.  Power user continue to have the option of specifying all key
values.