Use new kind of modal proxy since our version is deprecated
We will have to update the secret for the db url in prod (and the async version of it) to not use localhost anymore to flow traffic through the proxy. moreover we can delete the db_override_hack secret.

We may want to add more IPs to the list for outbound traffic to distribute load, but we only use 1 for now so we don't disrupt OTB gitlab support, which relies on a static outbound IP