PE-2708: Add 'remote' command to AutoDoc CLI
This change adds the 'remote' command to the local AutoDoc's CLI.  The
command provides the same functionality as the 'execute' command, but
all processing is done with Modal.  Additionally, this command accepts a
page ID for a page that can be pre-populated with sources rather than
using the sources specified in the TOML config file.