AutoTOML updates
This commit introduces the following changes:

* A new optional parameter has been added to the AutoTOML generate and
append commands.  This allows the user to specify a string that will be
injected into the user prompts when generating a greenfield TOML or
appending an existing TOML.  A typical use case would be to specify
how verbose the TOML output should be or limit how many sections are
generated/added.

* Several refinements to the prompts have been made to improve
consistency and respect the new user context.

* Perform 'substitutions' before appending and existing TOML