PE-2766: AutoDoc error from non-analyzable source
This commit prevents AutoDoc failures when the user specifies a source
that is not analyzable, and hence, has no derived content.  This case
leads to a condition where we attempt to build a sub-graph from a dag
for the offending file, but the file is not a node in the dag.

The solution is to simply detect the condition and then pass.  It is a
bit of a hack, but seems like the most simple option.

Ultimately, the front end should not allow tuning a codebase with
non-analyzable nodes.