PE-3020: Add logging to MCP server
The MCP server is currently set to mask error details meaning that
exception messages are not returned to the agent UNLESS they are
ToolErrors.  This exception type is used to report codebase that are
not supported or no content exists for a codebase.

Additionally, FastMCP ErrorHandlingMiddleware and LoggingMiddleware are
added to the MCP server to provide more detailed error messages and
logging automatically.