Remove modal volume usage
We likely have issues when an already running container can't see volume changes, leading to failed file loading in tech docs on occation.

We could refresh the volume before trying to download, but then we risk someone already writing to the volume

This issue can arise when reusing containers for tech docs concurrently across invocations of inspector.

Instead, we opt to put the symbol table in s3 and download it into each container once as needed per container