PE-3089: Steer chat from violating tuning
This is an attempt at preventing chat from violating tuning
restrictions. It adds a 'tuned' flag to the DataSource class, which is
visible to the agent via the DataSourceSystemMessage.  This simply
allows the agent to know if the sources are tuned or not.  Complementary
to this, the FileDocumentationTool and CodeMapTool description have bee
updated to nudge the agent to avoid using the tools if the sources are
tuned and the file or directory is not listed in the DataSource message.